requirements.txt
Flask==3.0.3
mongoengine==0.28.2
pymongo[srv]==4.8.0
dnspython==2.6.1
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.2
Werkzeug==3.0.3
gunicorn==22.0.0
